backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-docx
python-pptx
google-generativeai
python-dotenv
pydantic
psycopg2-binary